# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MetaRadar",
    "version": "5.1.0"
  },
  "paths": {
...
    "/api/v1/developments": {
      "get": {
        "tags": [
          "Registry"
        ],
        "summary": "Get Developments Registry",
        "description": "Retrieve indexed clinical/regulatory developments with asset and company details.",
        "operationId": "get_developments_registry_api_v1_developments_get",
        "parameters": [
          {
            "name": "disease",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disease"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DevelopmentSummary"
                  },
                  "title": "Response Get Developments Registry Api V1 Developments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ingestion/run": {
      "post": {
        "tags": [
          "Live Ingestion & Data Connectors"
        ],
        "summary": "Trigger Live Connector Ingestion",
        "description": "Fetches live biomedical data from public APIs into raw_signals_bronze and logs health status.",
        "operationId": "trigger_ingestion_run_api_v1_ingestion_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/IngestionRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Trigger Ingestion Run Api V1 Ingestion Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DevelopmentSummary": {
        "properties": {
          "development_id": {
            "type": "string",
            "format": "uuid",
            "title": "Development Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "disease": {
            "type": "string",
            "title": "Disease"
          },
          "current_stage": {
            "type": "string",
            "title": "Current Stage"
          },
          "asset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Name"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "signal_count": {
            "type": "integer",
            "title": "Signal Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "development_id",
          "title",
          "disease",
          "current_stage",
          "created_at",
          "updated_at"
        ],
        "title": "DevelopmentSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestionRunRequest": {
        "properties": {
          "connector_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connector Ids",
            "description": "Optional list of connector IDs to run: pubmed, clinical_trials, fda, ema, newsapi. If None, runs all."
          },
          "force_backfill": {
            "type": "boolean",
            "title": "Force Backfill",
            "description": "Whether to force backfill window replay instead of incremental rolling window.",
            "default": false
          }
        },
        "type": "object",
        "title": "IngestionRunRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}